+++ Sources/Sentry/SentryUIViewControllerSwizziling.m
@@ -3,6 +3,7 @@
 #import "SentryPerformanceTracker.h"
 #import "SentrySwizzle.h"
 #import "SentryUIViewControllerPerformanceTracker.h"
+#import <SentryDispatchQueueWrapper.h>
 #import <SentryInAppLogic.h>
 #import <SentryOptions.h>
 #import <UIViewController+Sentry.h>
@@ -16,43 +17,85 @@
 static SentryInAppLogic *inAppLogic;
 
 + (void)startWithOptions:(SentryOptions *)options
+           dispatchQueue:(SentryDispatchQueueWrapper *)dispatchQueue
 {
     inAppLogic = [[SentryInAppLogic alloc] initWithInAppIncludes:options.inAppIncludes
                                                    inAppExcludes:options.inAppExcludes];
-    [SentryUIViewControllerSwizziling swizzleViewControllerInits];
+
+    [SentryUIViewControllerSwizziling swizzleRootViewController];
+
+    [SentryUIViewControllerSwizziling
+        swizzleSubclassesOf:[UIViewController class]
+              dispatchQueue:dispatchQueue
+               swizzleBlock:^(Class class) {
+                   [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:class];
+               }];
+}
+
+/**
+ * To be able to test this we put the logic in a extra method.
+ *
+ * Swizzle all custom UIViewControllers. First, fetch all subclasses of UIViewController and then
+ * swizzle them. As there is no straightforward way to get all sub-classes in Objective-C, the code
+ * first retrieves all classes in the runtime, iterates over all classes, and checks for every class
+ * if UIViewController is a parent. Cause loading all classes can take a few milliseconds, do this
+ * on a background thread, which should be fine because the SDK swizzles the root view controller
+ * and its children above. After finding all subclasses of the UIViewController, this method
+ * swizzles them on the main thread. Swizzling the UIViewControllers on a background thread led to
+ * crashes, see GH-1366.
+ *
+ * Previously, the code intercepted the ViewController initializers with
+ * swizzling to swizzle the lifecycle methods. This approach led to UIViewControllers crashing when
+ * using a convenience initializer, see GH-1355. The error occurred because our swizzling logic adds
+ * the method to swizzle if the class doesn't implement it. It seems like adding an extra
+ * initializer causes problems with the rules for initialization in Swift, see
+ * https://docs.swift.org/swift-book/LanguageGuide/Initialization.html#ID216.
+ */
++ (void)swizzleSubclassesOf:(Class)parentClass
+              dispatchQueue:(SentryDispatchQueueWrapper *)dispatchQueue
+               swizzleBlock:(void (^)(Class))block
+{
+    [dispatchQueue dispatchAsyncWithBlock:^{
+        int numClasses = objc_getClassList(NULL, 0);
+        Class *classes = (__unsafe_unretained Class *)malloc(sizeof(Class) * numClasses);
+        numClasses = objc_getClassList(classes, numClasses);
+
+        if (numClasses <= 0) {
+            [SentryLog logWithMessage:@"No classes found when retrieving class list."
+                             andLevel:kSentryLevelError];
+        }
+
+        // Storing the actual classes in an NSArray would call initialize of the class, which we
+        // must avoid as we are on a background thread here and dealing with UIViewControllers,
+        // which assume they are running on the main thread. Therefore, we store the indexes instead
+        // so we can search for the subclasses on a background thread.
+        NSMutableArray *indexesToSwizzle = [NSMutableArray new];
+        for (NSInteger i = 0; i < numClasses; i++) {
+            Class superClass = classes[i];
+            do {
+                superClass = class_getSuperclass(superClass);
+            } while (superClass && superClass != parentClass);
+
+            if (superClass != nil) {
+                [indexesToSwizzle addObject:@(i)];
+            }
+        }
+
+        // We must swizzle the UIViewControllers on the main thread. Otherwise, we could crash.
+        [dispatchQueue dispatchOnMainQueue:^{
+            for (NSNumber *i in indexesToSwizzle) {
+                NSInteger index = [i integerValue];
+                block(classes[index]);
+            }
+            free(classes);
+        }];
+    }];
 }
 
 // SentrySwizzleInstanceMethod declaration shadows a local variable. The swizzling is working
 // fine and we accept this warning.
 #    pragma clang diagnostic push
 #    pragma clang diagnostic ignored "-Wshadow"
-
-/**
- * Swizzle the some init methods of the view controller,
- * so we can swizzle user view controller subclass on demand.
- */
-+ (void)swizzleViewControllerInits
-{
-    static const void *swizzleViewControllerInitWithCoder = &swizzleViewControllerInitWithCoder;
-    SEL coderSelector = NSSelectorFromString(@"initWithCoder:");
-    SentrySwizzleInstanceMethod(UIViewController.class, coderSelector, SentrySWReturnType(id),
-        SentrySWArguments(NSCoder * coder), SentrySWReplacement({
-            [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:[self class] isNib:NO];
-            return SentrySWCallOriginal(coder);
-        }),
-        SentrySwizzleModeOncePerClassAndSuperclasses, swizzleViewControllerInitWithCoder);
-
-    static const void *swizzleViewControllerInitWithNib = &swizzleViewControllerInitWithNib;
-    SEL nibSelector = NSSelectorFromString(@"initWithNibName:bundle:");
-    SentrySwizzleInstanceMethod(UIViewController.class, nibSelector, SentrySWReturnType(id),
-        SentrySWArguments(NSString * nibName, NSBundle * bundle), SentrySWReplacement({
-            [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:[self class] isNib:YES];
-            return SentrySWCallOriginal(nibName, bundle);
-        }),
-        SentrySwizzleModeOncePerClassAndSuperclasses, swizzleViewControllerInitWithNib);
-
-    [SentryUIViewControllerSwizziling swizzleRootViewController];
-}
 
 /**
  * If an app targets, for example, iOS 13 or lower, the UIKit inits the initial/root view controller
@@ -119,14 +162,12 @@ static SentryInAppLogic *inAppLogic;
             NSString *message = @"UIViewControllerSwizziling Calling swizzleRootViewController.";
             [SentryLog logWithMessage:message andLevel:kSentryLevelDebug];
 
-            // We don't now if it the UIViewController uses a nib or not.
-            [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:viewControllerClass
-                                                                      isNib:YES];
+            [SentryUIViewControllerSwizziling swizzleViewControllerSubClass:viewControllerClass];
         }
     }
 }
 
-+ (void)swizzleViewControllerSubClass:(Class)class isNib:(BOOL)isNib
++ (void)swizzleViewControllerSubClass:(Class)class
 {
     if (![SentryUIViewControllerSwizziling shouldSwizzleViewController:class])
         return;
@@ -134,7 +175,7 @@ static SentryInAppLogic *inAppLogic;
     // This are the five main functions related to UI creation in a view controller.
     // We are swizzling it to track anything that happens inside one of this functions.
     [SentryUIViewControllerSwizziling swizzleViewLayoutSubViews:class];
-    [SentryUIViewControllerSwizziling swizzleLoadView:class isNib:isNib];
+    [SentryUIViewControllerSwizziling swizzleLoadView:class];
     [SentryUIViewControllerSwizziling swizzleViewDidLoad:class];
     [SentryUIViewControllerSwizziling swizzleViewWillAppear:class];
     [SentryUIViewControllerSwizziling swizzleViewDidAppear:class];
@@ -156,18 +197,20 @@ static SentryInAppLogic *inAppLogic;
     return [inAppLogic isInApp:classImageName];
 }
 
-+ (void)swizzleLoadView:(Class)class isNib:(BOOL)isNib
++ (void)swizzleLoadView:(Class)class
 {
     // The UIViewController only searches for a nib file if you do not override the loadView method.
     // When swizzling the loadView of a custom UIViewController, the UIViewController doesn't search
     // for a nib file and doesn't load a view. This would lead to crashes as no view is loaded. As a
     // workaround, we skip swizzling the loadView and accept that the SKD doesn't create a span for
     // loadView if the UIViewController doesn't implement it.
-    if (isNib) {
+    SEL selector = NSSelectorFromString(@"loadView");
+    IMP viewControllerImp = class_getMethodImplementation([UIViewController class], selector);
+    IMP classLoadViewImp = class_getMethodImplementation(class, selector);
+    if (viewControllerImp == classLoadViewImp) {
         return;
     }
 
-    SEL selector = NSSelectorFromString(@"loadView");
     SentrySwizzleInstanceMethod(class, selector, SentrySWReturnType(void), SentrySWArguments(),
         SentrySWReplacement({
             [SentryUIViewControllerPerformanceTracker.shared