+++ Tests/SentryTests/SentryCrash/SentryStacktraceBuilderTests.swift
@@ -4,27 +4,34 @@ import XCTest
 class SentryStacktraceBuilderTests: XCTestCase {
     
     private class Fixture {
-        func getSut() -> SentryStacktraceBuilder {
-            SentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: [])))
+        let queue = DispatchQueue(label: "SentryStacktraceBuilderTests")
+
+        var sut: SentryStacktraceBuilder {
+            return SentryStacktraceBuilder(crashStackEntryMapper: SentryCrashStackEntryMapper(inAppLogic: SentryInAppLogic(inAppIncludes: [], inAppExcludes: [])))
         }
     }
-    
+
+    private var fixture: Fixture!
+
+    override func setUp() {
+        super.setUp()
+        fixture = Fixture()
+    }
+
     override func tearDown() {
         super.tearDown()
-        SentrySDK.close()
+        clearTestState()
     }
     
-    private let fixture = Fixture()
-    
     func testEnoughFrames() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread()
+        let actual = fixture.sut.buildStacktraceForCurrentThread()
         
         // The stacktrace has usually more than 40 frames. Feel free to change the number if the tests are failing
-        XCTAssertTrue(30 < actual.frames.count, "Not enough stacktrace frames.")
+        XCTAssertTrue(30 < actual.frames.count, "Not enough stacktrace frames. It should be more than 30, but was \(actual.frames.count)")
     }
     
     func testFramesAreFilled() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread()
+        let actual = fixture.sut.buildStacktraceForCurrentThread()
         
         // We don't know the actual values of the frames so we can't write
         // deterministic tests here. Therefore we just make sure they are
@@ -38,7 +45,7 @@ class SentryStacktraceBuilderTests: XCTestCase {
     }
     
     func testFramesDontContainBuilderFunction() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread()
+        let actual = fixture.sut.buildStacktraceForCurrentThread()
         
         let result = actual.frames.contains { frame in
             return frame.function?.contains("buildStacktraceForCurrentThread") ?? false
@@ -48,7 +55,7 @@ class SentryStacktraceBuilderTests: XCTestCase {
     }
     
     func testFramesOrder() {
-        let actual = fixture.getSut().buildStacktraceForCurrentThread()
+        let actual = fixture.sut.buildStacktraceForCurrentThread()
         
         // Make sure the first 4 frames contain main
         let frames = actual.frames[...3]
@@ -60,38 +67,40 @@ class SentryStacktraceBuilderTests: XCTestCase {
     }
     
     func testAsyncStacktraces() throws {
-        SentrySDK.start(options: ["dsn": TestConstants.dsnAsString(username: "SentrySDKTests")])
+        SentrySDK.start { options in
+            options.dsn = TestConstants.dsnAsString(username: "SentryStacktraceBuilderTests")
+            options.stitchAsyncCode = true
+        }
         
-        let group = DispatchGroup()
+        let expect = expectation(description: "testAsyncStacktraces")
 
-        DispatchQueue.main.async {
-            group.enter()
-            self.asyncFrame1(group: group)
+        fixture.queue.async {
+            self.asyncFrame1(expect: expect)
         }
         
-        group.waitWithTimeout()
+        wait(for: [expect], timeout: 2)
     }
     
-    func asyncFrame1(group: DispatchGroup) {
-        DispatchQueue.main.asyncAfter(deadline: DispatchTime.now()) {
-            self.asyncFrame2(group: group)
+    func asyncFrame1(expect: XCTestExpectation) {
+        fixture.queue.asyncAfter(deadline: DispatchTime.now()) {
+            self.asyncFrame2(expect: expect)
         }
     }
     
-    func asyncFrame2(group: DispatchGroup) {
-        DispatchQueue.main.async {
-            self.asyncAssertion(group: group)
+    func asyncFrame2(expect: XCTestExpectation) {
+        fixture.queue.async {
+            self.asyncAssertion(expect: expect)
         }
     }
     
-    func asyncAssertion(group: DispatchGroup) {
-        let actual = self.fixture.getSut().buildStacktraceForCurrentThread()
+    func asyncAssertion(expect: XCTestExpectation) {
+        let actual = fixture.sut.buildStacktraceForCurrentThread()
 
         let filteredFrames = actual.frames.filter { frame in
             return frame.function?.contains("testAsyncStacktraces") ?? false ||
-                frame.function?.contains("asyncFrame1") ?? false ||
-                frame.function?.contains("asyncFrame2") ?? false ||
-                frame.function?.contains("asyncAssertion") ?? false
+            frame.function?.contains("asyncFrame1") ?? false ||
+            frame.function?.contains("asyncFrame2") ?? false ||
+            frame.function?.contains("asyncAssertion") ?? false
         }
         let startFrames = actual.frames.filter { frame in
             return frame.stackStart?.boolValue ?? false
@@ -100,6 +109,6 @@ class SentryStacktraceBuilderTests: XCTestCase {
         XCTAssertTrue(filteredFrames.count >= 4, "The Stacktrace must include the async callers.")
         XCTAssertTrue(startFrames.count >= 3, "The Stacktrace must have async continuation markers.")
 
-        group.leave()
+        expect.fulfill()
     }
 }